// src/types.ts
export interface rumseyData {
  bounds: string;
  catalog_url: string;
  collection_id: string;
  contributor: string;
  created: string;
  creator: string;
  date: number;
  description: string;
  dpi: string;
  external_id: string;
  georeference_changed: string;
  georeference_is_finished: string;
  id: string;
  image_url: string;
  is_reviewed: string;
  link_url: string;
  physical_size: number[];
  pubdate: number;
  publisher: string;
  reviewed: string;
  scale: string;
  title: string;
  viewer_url: string;
  filename: number;
  displayName: string;
  mediaType: string;
  relayButtonUrl: string;
  relayButtonTitle: string;
  fieldValues: string;
  relatedFieldValues: string;
  urlSize4: string;
  urlSize2: string;
  urlSize3: string;
  iiifManifest: string;
  urlSize0: string;
  urlSize1: string;
  affine_transformation: string;
  changed: string;
  created_georeference: string;
  cutline: number[][];
  gcps: string;
  transformation_method: string;
  map_id: string;
  list_no: number;
  image_no: number;
}